Share the co-tenant subdomain scan between revoke and suspend
`active_addresses_on_subdomain` was copy-pasted byte-for-byte into
lambda/api/revoke/function.py and lambda/api/suspend_address/function.py.
Both handlers already import helper.py, so the single copy now lives
there next to the DNS-record helpers it guards, scanning the same
`cabal-addresses` table through helper's existing `ddb_table` handle.

Also corrects two stale symbol names in the revoke security comment:
`delete_dns_records` and `other_addresses_on_subdomain` no longer exist;
the real names are `delete_address_dns_records` and
`active_addresses_on_subdomain`.
