+++ lambda/api/revoke/function.py
@@ -4,6 +4,7 @@ import json
 import os
 from datetime import datetime, timezone
 import boto3  # pylint: disable=import-error
+from helper import active_addresses_on_subdomain  # pylint: disable=import-error
 from helper import delete_address_dns_records  # pylint: disable=import-error
 from helper import parse_json_body  # pylint: disable=import-error
 from helper import user_authorized_for_sender  # pylint: disable=import-error
@@ -34,10 +35,11 @@ def handler(event, _context):
     # Take subdomain/tld/zone from the STORED row for `address`, never from the
     # request body. Authorization above is on `address` only, so honoring a
     # client-supplied subdomain/tld would let a caller who owns any one address
-    # delete another user's DNS records: delete_dns_records targets
-    # `{subdomain}.{tld}`, and the co-tenant guard (other_addresses_on_subdomain)
-    # returns False for a single-tenant victim subdomain, so the DELETE would
-    # proceed. The caller owns `address`, so its row is the authoritative source.
+    # delete another user's DNS records: delete_address_dns_records targets
+    # `{subdomain}.{tld}`, and the co-tenant guard
+    # (active_addresses_on_subdomain) returns False for a single-tenant victim
+    # subdomain, so the DELETE would proceed. The caller owns `address`, so its
+    # row is the authoritative source.
     item = table.get_item(Key={'address': address}).get('Item') or {}
     subdomain = item.get('subdomain')
     tld = item.get('tld')
@@ -72,32 +74,6 @@ def handler(event, _context):
             'address': address
         })
     }
-
-
-def active_addresses_on_subdomain(subdomain, tld, address):
-    '''Checks if other non-suspended addresses share the same subdomain and TLD'''
-    scan_kwargs = {
-        'FilterExpression': (
-            'subdomain = :sub AND tld = :tld AND address <> :addr '
-            'AND (attribute_not_exists(#s) OR #s = :false)'
-        ),
-        'ExpressionAttributeNames': {'#s': 'suspended'},
-        'ExpressionAttributeValues': {
-            ':sub': subdomain,
-            ':tld': tld,
-            ':addr': address,
-            ':false': False
-        },
-        'ProjectionExpression': 'address'
-    }
-    while True:
-        response = table.scan(**scan_kwargs)
-        if response.get('Items'):
-            return True
-        if 'LastEvaluatedKey' not in response:
-            break
-        scan_kwargs['ExclusiveStartKey'] = response['LastEvaluatedKey']
-    return False
 
 
 def revoke_address(address):
